# docs/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SME Digitization API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sales": {
      "post": {
        "tags": [
          "sales"
        ],
        "summary": "Create Sale",
        "operationId": "create_sale_api_v1_sales_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sales"
        ],
        "summary": "List Sales",
        "operationId": "list_sales_api_v1_sales_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sales/{sale_id}": {
      "get": {
        "tags": [
          "sales"
        ],
        "summary": "Get Sale",
        "operationId": "get_sale_api_v1_sales__sale_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sale_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sale Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sales/{sale_id}/refund": {
      "post": {
        "tags": [
          "sales"
        ],
        "summary": "Refund Sale",
        "operationId": "refund_sale_api_v1_sales__sale_id__refund_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sale_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sale Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleRefundCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRefundOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/summary": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Summary Report",
        "operationId": "summary_report_api_v1_reports_summary_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/low-stock": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Low Stock Report",
        "operationId": "low_stock_report_api_v1_reports_low_stock_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LowStockReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/cashbook": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Cashbook Report",
        "operationId": "cashbook_report_api_v1_reports_cashbook_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashbookReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/top-products": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Top Products Report",
        "operationId": "top_products_report_api_v1_reports_top_products_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopProductsReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashbookReport": {
        "properties": {
          "cash_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cash Total"
          },
          "qr_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qr Total"
          },
          "bank_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Bank Total"
          },
          "credit_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credit Total"
          }
        },
        "type": "object",
        "required": [
          "cash_total",
          "qr_total",
          "bank_total",
          "credit_total"
        ],
        "title": "CashbookReport"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LowStockItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "stock_qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Stock Qty"
          },
          "low_stock_threshold": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Low Stock Threshold"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "stock_qty",
          "low_stock_threshold"
        ],
        "title": "LowStockItem"
      },
      "LowStockReport": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LowStockItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LowStockReport"
      },
...
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "CASH",
          "QR",
          "BANK",
          "CREDIT"
        ],
        "title": "PaymentMethod"
      },
...
      "SaleCreate": {
        "properties": {
          "sale_type": {
            "$ref": "#/components/schemas/SaleType"
          },
          "payment_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SaleItemCreate"
            },
            "type": "array",
            "title": "Items"
          },
          "payments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SalePaymentCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payments"
          }
        },
        "type": "object",
        "required": [
          "sale_type",
          "items"
        ],
        "title": "SaleCreate"
      },
      "SaleItemCreate": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Qty"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Unit Price"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "qty",
          "unit_price"
        ],
        "title": "SaleItemCreate"
      },
      "SaleItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sale_id": {
            "type": "string",
            "title": "Sale Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price"
          },
          "line_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Line Total"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sale_id",
          "product_id",
          "qty",
          "unit_price",
          "line_total"
        ],
        "title": "SaleItemOut"
      },
      "SaleListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SaleOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "SaleListResponse"
      },
      "SaleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "sale_type": {
            "type": "string",
            "title": "Sale Type"
          },
          "payment_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Method"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SaleItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "payments": {
            "items": {
              "$ref": "#/components/schemas/SalePaymentOut"
            },
            "type": "array",
            "title": "Payments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "store_id",
          "sale_type",
          "customer_id",
          "total_amount",
          "created_at",
          "updated_at",
          "items"
        ],
        "title": "SaleOut"
      },
      "SalePaymentCreate": {
        "properties": {
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "method",
          "amount"
        ],
        "title": "SalePaymentCreate"
      },
      "SalePaymentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sale_id": {
            "type": "string",
            "title": "Sale Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sale_id",
          "method",
          "amount",
          "created_at"
        ],
        "title": "SalePaymentOut"
      },
      "SaleRefundCreate": {
        "properties": {
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SaleRefundItemCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "SaleRefundCreate"
      },
      "SaleRefundItemCreate": {
        "properties": {
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "sale_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sale Item Id"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "qty"
        ],
        "title": "SaleRefundItemCreate"
      },
      "SaleRefundItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "refund_id": {
            "type": "string",
            "title": "Refund Id"
          },
          "sale_id": {
            "type": "string",
            "title": "Sale Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price"
          },
          "line_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Line Total"
          }
        },
        "type": "object",
        "required": [
          "id",
          "refund_id",
          "sale_id",
          "product_id",
          "qty",
          "unit_price",
          "line_total"
        ],
        "title": "SaleRefundItemOut"
      },
      "SaleRefundOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "sale_id": {
            "type": "string",
            "title": "Sale Id"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SaleRefundItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "store_id",
          "sale_id",
          "amount",
          "reason",
          "created_at",
          "items"
        ],
        "title": "SaleRefundOut"
      },
      "SaleType": {
        "type": "string",
        "enum": [
          "CASH",
          "CREDIT",
          "MIXED"
        ],
        "title": "SaleType"
      },
...
      "SummaryReport": {
        "properties": {
          "total_sales": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Sales"
          },
          "total_expenses": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Expenses"
          },
          "estimated_profit": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Estimated Profit"
          },
          "credit_outstanding": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credit Outstanding"
          }
        },
        "type": "object",
        "required": [
          "total_sales",
          "total_expenses",
          "estimated_profit",
          "credit_outstanding"
        ],
        "title": "SummaryReport"
      },
...
      "TopProductItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "qty_sold": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty Sold"
          },
          "revenue": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Revenue"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "qty_sold",
          "revenue"
        ],
        "title": "TopProductItem"
      },
      "TopProductsReport": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TopProductItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TopProductsReport"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}